Add useTimer custom hook
